feat: msg to separate file
