+++ artifacts/checksums.txt
@@ -1,4 +1,4 @@
 09f25ef4c60bd3135bef6e3399033f0563243d11819d0ea79561506e102819ec  ibc_transfer.wasm
 6109ae928c24b7dd861b9eec8f78edc46f8af984b98719923f432f946b64cdd5  neutron_interchain_queries.wasm
-67bfd93652b57742fcc43933bd8caa0c1095fc1706a47144ee1e8559322ab1ad  neutron_interchain_txs.wasm
+a9bec1e5e68877931f827eb2a68cac6f12a27df8dda3bde06f8bcd1c616af28f  neutron_interchain_txs.wasm
 cf66ca7c6f7027538671f1c0ea547f35f71dbfcc50907bc0bfb41a0782950dab  reflect.wasm
